Created Site settings screen and tile
1. Created a new sites tile in this hierarchy
     Settings -> Catalogs -> Sites
2. Created a new sites screen which was previously under Events
