+++ lib/screens/settings/collevent_settings.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:nahpu/services/providers/collevents.dart';
-import 'package:nahpu/services/providers/sites.dart';
 import 'package:nahpu/screens/settings/common.dart';
 import 'package:nahpu/screens/shared/common.dart';
 import 'package:nahpu/services/collevent_services.dart';
-import 'package:nahpu/services/site_services.dart';
 
 class CollEventSelection extends StatefulWidget {
   const CollEventSelection({super.key});
@@ -26,7 +24,6 @@ class _CollEventSelectionState extends State<CollEventSelection> {
           sections: [
             CollMethodSettings(),
             PersonnelRoleSetting(),
-            HabitatTypeSettings(),
           ],
         ),
       ),
@@ -107,44 +104,6 @@ class PersonnelRoleSetting extends ConsumerWidget {
       resetLabel: 'Match database roles',
       onReset: () {
         CollEvenPersonnelServices(ref: ref).getAllRoles();
-      },
-    );
-  }
-}
-
-class HabitatTypeSettings extends ConsumerWidget {
-  const HabitatTypeSettings({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    TextEditingController controller = TextEditingController();
-    return SettingChips(
-      title: 'Habitat types',
-      controller: controller,
-      chipList: ref.watch(habitatTypeProvider).when(
-            data: (data) {
-              return data.map((e) {
-                return CommonSettingChip(
-                  text: e,
-                  primaryColor: Theme.of(context).colorScheme.tertiary,
-                  onDeleted: () {
-                    HabitatServices(ref: ref).removeHabitat(e);
-                  },
-                );
-              }).toList();
-            },
-            loading: () => [const CommonProgressIndicator()],
-            error: (e, _) => [Text('Error: $e')],
-          ),
-      labelText: 'Add habitat type',
-      hintText: 'Enter habitat type',
-      onPressed: () {
-        HabitatServices(ref: ref).addHabitat(controller.text.trim());
-        controller.clear();
-      },
-      resetLabel: 'Match database habitat types',
-      onReset: () {
-        HabitatServices(ref: ref).getAllHabitats();
       },
     );
   }
