+++ lib/screens/settings/settings.dart
@@ -6,6 +6,7 @@ import 'package:nahpu/screens/projects/personnel/manage_personnel.dart';
 import 'package:nahpu/screens/projects/taxonomy/taxon_list.dart';
 import 'package:nahpu/screens/settings/catalog_format.dart';
 import 'package:nahpu/screens/settings/collevent_settings.dart';
+import 'package:nahpu/screens/settings/site_settings.dart';
 import 'package:nahpu/screens/settings/common.dart';
 import 'package:nahpu/screens/settings/db_settings.dart';
 import 'package:nahpu/screens/shared/common.dart';
@@ -62,6 +63,18 @@ class CatalogSettings extends StatelessWidget {
       children: [
         CatalogFmtSection(selectedFmt: matchCatFmtToTaxonGroup(catalogFmt)),
         const CollEventSection(),
+        CommonSettingTile(
+          title: 'Sites',
+          icon: Icons.place_outlined,
+          label: 'Edit habitat types and site metadata',
+          isNavigation: true,
+          onTap: () => Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => const SiteSelection(),
+            ),
+          ),
+        ),
         SpecimenSection(catalogFmt: catalogFmt),
       ],
     );
